server/backendapi/requirements.txt
Django==4.2.0
social-auth-app-django==5.2.0
python-dotenv==1.0.0
django-allauth==0.54.0
